refactor/filter: Move filter utilities into utils.search_utils.filters
